+++ test/PHPMailer/PHPMailerTest.php
@@ -1258,16 +1258,6 @@ EOT;
         $this->Mail->smtpClose();
     }
 
-    public function testGivenIdnAddress_addAddress_returns_true()
-    {
-        if (file_exists(\PHPMAILER_INCLUDE_DIR . '/test/fakefunctions.php') === false) {
-            $this->markTestSkipped('/test/fakefunctions.php file not found');
-        }
-
-        include \PHPMAILER_INCLUDE_DIR . '/test/fakefunctions.php';
-        $this->assertTrue($this->Mail->addAddress('test@françois.ch'));
-    }
-
     public function testErroneousAddress_addAddress_returns_false()
     {
         $this->assertFalse($this->Mail->addAddress('mehome.com'));
